Extract PropertyFilterService from listener.
